Get the script working on linux (without dnsmasq)

+++ setup-docker.sh
@@ -1,21 +1,27 @@
+#!/usr/bin/env bash
 TLD=test
+if [ "$(uname)" == "Darwin" ]; then
+  DNSPORT=5300
+  sudo mkdir -p /etc/resolver
+  sudo rm /etc/resolver/${TLD}
+  sudo touch /etc/resolver/${TLD}
+  echo 'nameserver 127.0.0.1' | sudo tee -a /etc/resolver/${TLD} > /dev/null
+  echo 'port 5300' | sudo tee -a /etc/resolver/${TLD} > /dev/null
 
-sudo mkdir -p /etc/resolver
-sudo rm /etc/resolver/${TLD}
-sudo touch /etc/resolver/${TLD}
-echo 'nameserver 127.0.0.1' | sudo tee -a /etc/resolver/${TLD} > /dev/null
-echo 'port 5300' | sudo tee -a /etc/resolver/${TLD} > /dev/null
 
-# start dnsmasq
-docker stop dnsmasq || true && docker rm dnsmasq || true
-docker run -d \
-   --name dnsmasq \
-   --restart always \
-   -p 5300:53/tcp \
-   -p 5300:53/udp \
-   --cap-add NET_ADMIN \
-   andyshinn/dnsmasq \
-   --address=/${TLD}/127.0.0.1
+  # start dnsmasq
+  docker stop dnsmasq || true && docker rm dnsmasq || true
+  docker run -d \
+     --name dnsmasq \
+     --restart always \
+     -p $DNSPORT:53/tcp \
+     -p $DNSPORT:53/udp \
+     --cap-add NET_ADMIN \
+     andyshinn/dnsmasq \
+     --address=/${TLD}/127.0.0.1
+else
+  echo "Please add your dev hosts to /etc/hosts"
+fi
 
 # start haproxy
 docker stop haproxy || true && docker rm haproxy || true
